Backend/grn_bazar/requirements.txt
asgiref==3.7.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
branca==0.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
Choco==1.0.5
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.7.1
distlib==0.3.7
distro==1.9.0
Django==4.2.4
django-autoslug==1.9.9
django-cors-headers==4.3.0
django-filter==23.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
filelock==3.12.3
folium==0.15.0
h11==0.14.0
html5lib==1.1
httpie==3.2.2
idna==3.4
Jinja2==3.1.2
lxml==5.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
mysqlclient==2.2.1
numpy==1.26.2
opencage==2.3.1
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
phonenumbers==8.13.25
pika==1.3.2
Pillow==10.1.0
platformdirs==3.10.0
psycopg2-binary==2.9.7
py4j==0.10.9.7
pycparser==2.21
Pygments==2.17.2
PyGreSQL==5.2.5
PyJWT==1.7.1
pyOpenSSL==23.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
selenium==4.0.0b4
setuptools==70.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tabula-py==2.9.0
tk==0.1.0
trio==0.23.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==1.26.18
urllib3-secure-extra==0.1.0
virtualenv==20.24.3
webencodings==0.5.1
wsproto==1.2.0
